Fleet State
** Brought to you by Hack Week **

This was a feature in the 1.0 days where an operator
could find insight into the current state of all
their deployed servers.

After the inception of 2.0 we lost this insight as the 1.0
builders were gone and Nomad didn't give us the data.

Using the limited functionality of Nomad we can scrape the
current state of the fleet one host at a time. This takes
a while, but it seems like the safest option. In a scenario
where there are hundreds of nomad clients, I don't know how
long this will take to run, but it seemed safer than forking
hundreds of `kubectl` commands simultaneously.

The VM service gives us some insight via an exposed endpoint
where curl is already installed. Using this data we can display,
in a human readable fashion, how many machines of each type are
currently deployed.

Runner support is coming soon after the endpoint is published.